pyproject.toml
[tool.poetry]
name = "bitcoin_scalper"
version = "0.1.0"
description = "Bot de trading algorithmique BTC/USD avec pipeline ML, gestion du risque, et supervision."
authors = ["Auteur <email@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
numpy = "1.26.4"
pandas = "*"
scikit-learn = "*"
xgboost = "*"
torch = "*"
requests = "*"
fastapi = "*"
uvicorn = "*"
joblib = "*"
matplotlib = "*"
seaborn = "*"
psycopg2-binary = "*"
pyotp = "*"
prometheus_client = "*"
cryptography = "^45.0.2"
ta = "^0.11.0"
python-multipart = "*"
# MetaTrader5 = "*" # Commenté car non compatible macOS et non nécessaire pour les tests actuels
shap = "*"
optuna = "^4.3.0"
sphinx = "^8.2.3"
dvc = "^3.59.2"
lightgbm = "^4.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
ruff = "*"
httpx = "*"

[tool.poetry.scripts]
# lint = "ruff ." # Ligne supprimée car non supportée par Poetry
# test = "pytest" # Ligne supprimée car non supportée par Poetry
# coverage = "pytest --cov=app --cov=bot --cov=windows --cov-report=xml --cov-report=term-missing"  # Ligne supprimée car non supportée par Poetry

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=app/core --cov=bot/connectors --cov-report=term-missing --cov-fail-under=95"
pythonpath = "." 